refactor(hud): split background service worker into cohesive modules
