+++ src/background.ts
@@ -5,43 +5,18 @@
  * @see https://zenn.dev/dotdotdot/articles/b123e67552fe3c
  */
 import PokerChaseService, {
-  ApiEvent,
-  ApiType,
   BATTLE_TYPE_FILTERS,
-  PlayerStats,
-  PokerChaseDB,
-  ApiMessage,
-  validateApiEvent,
-  isApiEventType,
-  parseApiEvent,
-  getValidationError,
-  isApplicationApiEvent
+  PokerChaseDB
 } from './app'
-import { EntityConverter } from './entity-converter'
-import { saveEntities, findLatestPlayerDealEvent } from './utils/database-utils'
-import type { AllPlayersRealTimeStats } from './realtime-stats/realtime-stats-service'
-import type { Options } from './components/Popup'
-import type { HandLogEvent } from './types/hand-log'
 import { defaultRegistry, defaultStatDisplayConfigs, mergeStatDisplayConfigs } from './stats'
 import type { StatsRegistry } from './stats/registry'
-import type {
-  ChromeMessage,
-  ExportProgressMessage,
-  HandLogEventMessage,
-  ImportProgressMessage,
-  ImportStatusMessage,
-  LatestStatsMessage,
-  MessageResponse,
-  RebuildProgressMessage
-} from './types/messages'
 import { firebaseAuthService } from './services/firebase-auth-service'
 import { autoSyncService } from './services/auto-sync-service'
 import { content_scripts } from '../manifest.json'
-import { DATABASE_CONSTANTS } from './constants/database'
+import { registerStreamSubscriptions } from './background/ports'
+import { registerEventIngestion } from './background/event-ingestion'
+import { registerMessageRouter } from './background/message-router'
 /** !!! CONTENT_SCRIPTS、WEB_ACCESSIBLE_RESOURCESからインポートしないこと !!! */
-
-const PING_INTERVAL_MS = 10 * 1000
-const IMPORT_CHUNK_SIZE = DATABASE_CONSTANTS.IMPORT_CHUNK_SIZE
 
 // Get game URL pattern from manifest
 const gameUrlPattern = content_scripts[0]!.matches[0]!
@@ -76,55 +51,6 @@ service.ready.then(async () => {
 }).catch(err => {
   console.error('[background] PokerChaseService initialization failed:', err)
 })
-
-interface ImportSession {
-  chunks: string[]
-  totalChunks: number
-  fileName: string
-}
-let currentImportSession: ImportSession | null = null
-
-interface OperationState {
-  type: 'idle' | 'export' | 'import' | 'rebuild'
-  format?: 'json' | 'pokerstars'
-  progress?: number
-  processed?: number
-  total?: number
-  message?: string
-}
-
-let currentOperationState: OperationState = { type: 'idle' }
-
-let lastKnownStats: PlayerStats[] = []
-
-// Store latest real-time stats (module scope: shared across all port connections)
-let latestRealTimeStats: AllPlayersRealTimeStats | undefined
-
-/**
- * `chrome.runtime.onConnect`で接続されたポートの集合
- * ストリームイベントをブロードキャストする際の送信先として利用する
- */
-const connectedPorts = new Set<chrome.runtime.Port>()
-
-/**
- * 接続中の全ポートにメッセージをブロードキャストする
- * 切断済みポートを検出した場合は`connectedPorts`から取り除く
- */
-const broadcastMessage = (data: { stats: PlayerStats[], evtDeal?: ApiEvent<ApiType.EVT_DEAL>, realTimeStats?: AllPlayersRealTimeStats } | string) => {
-  connectedPorts.forEach(port => {
-    try {
-      port.postMessage(data)
-    } catch (error: unknown) {
-      if (error instanceof Error) {
-        /** when `content_script` is inactive */
-        if (error.message === 'Attempting to use a disconnected port object')
-          connectedPorts.delete(port)
-        else
-          console.error(error)
-      }
-    }
-  })
-}
 
 /** 拡張更新時の処理 */
 chrome.runtime.onInstalled.addListener(async details => {
@@ -174,1147 +100,14 @@ chrome.storage.sync.get('options', (result: Record<string, any>) => {
   }
 })
 
-
 /**
  * データエクスポート機能
  */
-chrome.runtime.onMessage.addListener((request: ChromeMessage, sender: chrome.runtime.MessageSender, sendResponse: (response: MessageResponse) => void) => {
-  if (request.action === 'exportData') {
-    // Block concurrent operations
-    if (currentOperationState.type !== 'idle') {
-      console.warn(`[exportData] Blocked: operation already in progress (${currentOperationState.type})`)
-      sendResponse({ success: false, error: '別の処理が実行中です' })
-      return true
-    }
-    exportData(request.format)
-      .then(() => sendResponse({ success: true }))
-      .catch(error => {
-        console.error('Export error:', error)
-        sendResponse({ success: false, error: error.message })
-      })
-    return true // 非同期レスポンスを示す
-  } else if (request.action === 'importData') {
-    importData(request.data)
-      .then((result) => {
-        chrome.runtime.sendMessage<ImportStatusMessage>({
-          action: 'importStatus',
-          status: `インポートが完了しました (${result.successCount.toLocaleString()}件のログ${result.duplicateCount > 0 ? `, ${result.duplicateCount.toLocaleString()}件の重複をスキップ` : ''})`
-        })
-        sendResponse({ success: true })
-      })
-      .catch(error => {
-        console.error('Import error:', error)
-        chrome.runtime.sendMessage<ImportStatusMessage>({
-          action: 'importStatus',
-          status: 'インポートに失敗しました: ' + error.message
-        })
-        sendResponse({ success: false, error: error.message })
-      })
-    return true // 非同期レスポンスを示す
-  } else if (request.action === 'importDataInit') {
-    currentImportSession = {
-      chunks: [],
-      totalChunks: request.totalChunks,
-      fileName: request.fileName
-    }
-    sendResponse({ success: true })
-    return true
-  } else if (request.action === 'importDataChunk') {
-    if (!currentImportSession) {
-      sendResponse({ success: false, error: 'No import session active' })
-      return true
-    }
+registerMessageRouter(service, db, gameUrlPattern)
 
-    currentImportSession.chunks[request.chunkIndex] = request.chunkData
+registerStreamSubscriptions(service, gameUrlPattern)
 
-    sendResponse({ success: true })
-    return true
-  } else if (request.action === 'importDataProcess') {
-    if (!currentImportSession || currentImportSession.chunks.length !== currentImportSession.totalChunks) {
-      sendResponse({ success: false, error: 'Import session incomplete' })
-      return true
-    }
-
-    // すべてのチャンクを結合
-    const completeData = currentImportSession.chunks.join('')
-    currentImportSession = null // セッションをクリア
-
-    // データを処理
-    importData(completeData)
-      .then((result) => {
-        chrome.runtime.sendMessage<ImportStatusMessage>({
-          action: 'importStatus',
-          status: `インポートが完了しました (${result.successCount.toLocaleString()}件のログ${result.duplicateCount > 0 ? `, ${result.duplicateCount.toLocaleString()}件の重複をスキップ` : ''})`
-        })
-        sendResponse({ success: true })
-      })
-      .catch(error => {
-        console.error('Import error:', error)
-        chrome.runtime.sendMessage<ImportStatusMessage>({
-          action: 'importStatus',
-          status: 'インポートに失敗しました: ' + error.message
-        })
-        sendResponse({ success: false, error: error.message })
-      })
-    return true
-  } else if (request.action === 'updateBattleTypeFilter') {
-    // フィルター値がundefinedかチェック
-    if (!request.filterOptions) {
-      sendResponse({ success: false, error: 'No filter options provided' })
-      return true
-    }
-
-    // サービス内のフィルターを更新
-    service.setBattleTypeFilter(request.filterOptions)
-      .then(() => {
-        sendResponse({ success: true })
-      })
-      .catch((error: Error) => {
-        console.error('[background.ts] Filter update error:', error)
-        sendResponse({ success: false, error: error.message })
-      })
-
-    // ストレージに保存
-    const storageUpdate: Partial<Options> = {
-      filterOptions: request.filterOptions
-    }
-    chrome.storage.sync.set({ options: storageUpdate })
-
-    // コンテンツスクリプトにメッセージを転送
-    chrome.tabs.query({ url: gameUrlPattern }, tabs => {
-      tabs.forEach(tab => tab.id && chrome.tabs.sendMessage(tab.id, request))
-    })
-
-    // 新しいフィルターに基づいてHUD表示を強制更新
-    if (lastKnownStats.length > 0) {
-      // 現在の席ユーザーIDで計算を再トリガー
-      service.statsOutputStream.write(lastKnownStats.map(stat => stat.playerId))
-    }
-
-    return true // 非同期レスポンスを示す
-  } else if (request.action === 'requestLatestStats') {
-    getLatestSessionStats()
-      .then(stats => {
-        if (sender.tab?.id) {
-          chrome.tabs.sendMessage<LatestStatsMessage>(sender.tab.id, {
-            action: 'latestStats',
-            stats: stats
-          })
-        }
-        sendResponse({ success: true })
-      })
-      .catch(error => {
-        console.error('Error getting latest stats:', error)
-        sendResponse({ success: false, error: error.message })
-      })
-    return true
-  } else if (request.action === 'deleteAllData') {
-    // ログと設定を含むすべてのデータを削除
-    deleteAllData()
-      .then(() => {
-        sendResponse({ success: true })
-        // キャッシュされた統計をクリア
-        lastKnownStats = []
-      })
-      .catch(error => {
-        console.error('Error deleting data:', error)
-        sendResponse({ success: false, error: error.message })
-      })
-    return true
-  } else if (request.action === 'firebaseAuthStatus') {
-    // Check current auth status
-    const isSignedIn = firebaseAuthService.isSignedIn()
-    const userInfo = firebaseAuthService.getUserInfo()
-
-    sendResponse({ success: true, isSignedIn, userInfo })
-    return true
-  } else if (request.action === 'firebaseSignIn') {
-    // Firebase sign in
-    handleFirebaseSignIn()
-      .then(() => sendResponse({ success: true }))
-      .catch(error => {
-        console.error('Firebase sign in error:', error)
-        sendResponse({ success: false, error: error.message })
-      })
-    return true
-  } else if (request.action === 'firebaseSignOut') {
-    // Firebase sign out
-    handleFirebaseSignOut()
-      .then(() => sendResponse({ success: true }))
-      .catch(error => {
-        console.error('Firebase sign out error:', error)
-        sendResponse({ success: false, error: error.message })
-      })
-    return true
-  } else if (request.action === 'firebaseSyncToCloud' || request.action === 'firebaseSyncFromCloud') {
-    // Manual sync now uses auto sync service
-    autoSyncService.performSync()
-      .then(() => sendResponse({ success: true }))
-      .catch(error => {
-        console.error('Manual sync error:', error)
-        sendResponse({ success: false, error: error.message })
-      })
-    return true
-  } else if (request.action === 'manualSyncUpload') {
-    // Manual upload to cloud
-    autoSyncService.performSync('upload')
-      .then(() => sendResponse({ success: true }))
-      .catch(error => {
-        console.error('Manual upload error:', error)
-        sendResponse({ success: false, error: error.message })
-      })
-    return true
-  } else if (request.action === 'manualSyncDownload') {
-    // Manual download from cloud
-    autoSyncService.performSync('download')
-      .then(() => sendResponse({ success: true }))
-      .catch(error => {
-        console.error('Manual download error:', error)
-        sendResponse({ success: false, error: error.message })
-      })
-    return true
-  } else if (request.action === 'getSyncState') {
-    // Get current sync state
-    const state = autoSyncService.getSyncState()
-    sendResponse({ success: true, syncState: state })
-    return false
-  } else if (request.action === 'getUnsyncedCount') {
-    // Get unsynced event count
-    autoSyncService.getUnsyncedEventCount()
-      .then(count => {
-        sendResponse({ success: true, count })
-      })
-      .catch(error => {
-        console.error('Error getting unsynced count:', error)
-        sendResponse({ success: false, error: error.message })
-      })
-    return true
-  } else if (request.action === 'getSyncInfo') {
-    // Get detailed sync information
-    autoSyncService.getSyncInfo()
-      .then(info => {
-        sendResponse({ success: true, syncInfo: info })
-      })
-      .catch(error => {
-        console.error('Error getting sync info:', error)
-        sendResponse({ success: false, error: error.message })
-      })
-    return true
-  } else if (request.action === 'rebuildData') {
-    // Block concurrent operations
-    if (currentOperationState.type !== 'idle') {
-      console.warn(`[rebuildData] Blocked: operation already in progress (${currentOperationState.type})`)
-      sendResponse({ success: false, error: '別の処理が実行中です' })
-      return true
-    }
-    // 手動でのデータ再構築
-    console.log('[rebuildData] Starting manual data rebuild...')
-
-    // バッチモードで全データを再構築（ダウンロード同期と同じ処理）
-    rebuildAllData()
-      .then(() => {
-        console.log('[rebuildData] Data rebuild completed')
-        sendResponse({ success: true })
-      })
-      .catch(error => {
-        console.error('[rebuildData] Error rebuilding data:', error)
-        sendResponse({ success: false, error: error.message })
-      })
-    return true
-  } else if (request.action === 'getOperationState') {
-    console.log('[getOperationState]', JSON.stringify(currentOperationState))
-    sendResponse({ success: true, operationState: currentOperationState })
-    return true
-  }
-  return false
-})
-
-const exportData = async (format: string) => {
-  if (format === 'json') {
-    await exportJsonData(db)
-  } else if (format === 'pokerstars') {
-    await exportPokerStarsData()
-  }
-}
-
-/**
- * Import data from JSONL file
- * @param jsonlData JSONL string containing API events (one JSON object per line)
- * @returns Object containing import statistics
- */
-const importData = async (jsonlData: string): Promise<{ successCount: number, totalLines: number, duplicateCount: number }> => {
-  try {
-    currentOperationState = { type: 'import', progress: 0 }
-    console.log('[importData] Starting optimized import process with direct entity generation')
-    const startTime = performance.now()
-
-    // 既存キーを一括取得（最適化ポイント1）
-    console.log('[importData] Loading existing keys...')
-    const existingKeys = new Set<string>()
-    await db.apiEvents
-      .orderBy('[timestamp+ApiTypeId]')
-      .keys(keys => {
-        keys.forEach(key => {
-          if (Array.isArray(key) && key.length === 2) {
-            existingKeys.add(`${key[0]}-${key[1]}`)
-          }
-        })
-      })
-    console.log(`[importData] Loaded ${existingKeys.size} existing keys`)
-
-    // 行で分割し、空行をフィルタリング
-    const lines = jsonlData.split('\n').filter(line => line.trim())
-    console.log(`[importData] Processing ${lines.length} lines`)
-
-    // バッチモードを有効化
-    service.setBatchMode(true)
-
-    // 直接エンティティ生成用のイベントを収集
-    const allNewEvents: ApiEvent[] = []
-
-    // メモリ問題を避けるためチャンク単位で処理
-    let processed = 0
-    let successCount = 0
-    let duplicateCount = 0
-    const errors: string[] = []
-
-    for (let i = 0; i < lines.length; i += IMPORT_CHUNK_SIZE) {
-      const chunkLines = lines.slice(i, i + IMPORT_CHUNK_SIZE)
-      const newEvents: ApiEvent[] = []
-
-      // チャンク内の各行をパース
-      for (let j = 0; j < chunkLines.length; j++) {
-        const lineNumber = i + j + 1
-        const line = chunkLines[j]
-        if (!line) continue
-
-        try {
-          const parsed = JSON.parse(line)
-
-          // タイムスタンプチェック（インポートでは必須）
-          if (!('timestamp' in parsed && parsed.timestamp)) {
-            errors.push(`Line ${lineNumber}: Missing timestamp`)
-            continue
-          }
-
-          // Zodスキーマ検証
-          const event = parseApiEvent(parsed)
-          if (!event) {
-            const result = validateApiEvent(parsed)
-            const errorDetails = result.error ? getValidationError(result.error)[0] : null
-            errors.push(`Line ${lineNumber}: ${errorDetails?.message || 'Validation failed'}`)
-            continue
-          }
-
-          // アプリケーション用のイベントかチェック
-          if (!isApplicationApiEvent(event)) {
-            // アプリケーションで使用しないApiTypeIdのイベントをスキップ
-            continue
-          }
-
-          const key = `${event.timestamp}-${event.ApiTypeId}`
-
-          // メモリ内で重複チェック（最適化ポイント2）
-          if (!existingKeys.has(key)) {
-            newEvents.push(event)
-            existingKeys.add(key) // 次の重複チェック用
-          } else {
-            duplicateCount++
-          }
-        } catch (parseError) {
-          // 無効なJSON行をスキップ
-          if (line.trim()) {
-            errors.push(`Line ${lineNumber}: Invalid JSON`)
-          }
-        }
-      }
-
-      // 新規イベントをbulkAddで一括保存（最適化ポイント3）
-      if (newEvents.length > 0) {
-        try {
-          await db.apiEvents.bulkAdd(newEvents)
-          successCount += newEvents.length
-          allNewEvents.push(...newEvents)
-        } catch (dbError) {
-          // 部分的な失敗の場合、個別に保存を試みる
-          console.warn(`Bulk add failed for chunk ${Math.floor(i / IMPORT_CHUNK_SIZE) + 1}, falling back to individual adds:`, dbError)
-
-          for (const event of newEvents) {
-            try {
-              await db.apiEvents.add(event)
-              successCount++
-              allNewEvents.push(event)
-            } catch (individualError) {
-              // 個別エラーは重複以外の場合のみログ
-              const errorMessage = individualError instanceof Error ? individualError.message : String(individualError)
-              if (!errorMessage.includes('Key already exists')) {
-                errors.push(`Event at timestamp ${event.timestamp}: ${errorMessage}`)
-              } else {
-                duplicateCount++
-                successCount-- // 重複の場合は成功数から除外
-              }
-            }
-          }
-        }
-      }
-
-      processed += chunkLines.length
-
-      // Send progress update
-      const progress = Math.round((processed / lines.length) * 100)
-      currentOperationState = { type: 'import', progress, processed, total: lines.length }
-      chrome.runtime.sendMessage<ImportProgressMessage>({
-        action: 'importProgress',
-        progress: progress,
-        processed: processed,
-        total: lines.length,
-        duplicates: duplicateCount,
-        imported: successCount
-      })
-
-      // Log progress every 10%
-      if (progress % 10 === 0) {
-        console.log(`[importData] Progress: ${progress}% (${processed}/${lines.length} lines)`)
-      }
-
-      // Allow browser to breathe between chunks
-      await new Promise(resolve => setTimeout(resolve, 5))
-    }
-
-    if (errors.length > 0) {
-      console.warn(`[importData] Failed to import ${errors.length} lines (${((errors.length / lines.length) * 100).toFixed(2)}%)`)
-      if (errors.length <= 10) {
-        console.warn('Errors:', errors)
-      } else {
-        console.warn('First 10 errors:', errors.slice(0, 10))
-      }
-    }
-
-    const importTime = ((performance.now() - startTime) / 1000).toFixed(2)
-    console.log(`[importData] Import completed in ${importTime}s - Success: ${successCount}, Duplicates: ${duplicateCount}`)
-
-    // 直接エンティティ生成（Phase 2最適化）
-    if (allNewEvents.length > 0) {
-      console.log(`[importData] Generating entities from ${allNewEvents.length} new events...`)
-      const entityStartTime = performance.now()
-
-      try {
-        // EntityConverterを使用してエンティティを生成
-        const converter = new EntityConverter(service.session)
-        const entities = converter.convertEventsToEntities(allNewEvents)
-
-        console.log(`[importData] Generated entities - Hands: ${entities.hands.length}, Phases: ${entities.phases.length}, Actions: ${entities.actions.length}`)
-
-        // Save entities using common utility
-        await saveEntities(db, entities, {
-          onProgress: (counts) => {
-            console.log(`[importData] Saved/updated ${counts.hands} hands, ${counts.phases} phases, ${counts.actions} actions`)
-          }
-        })
-
-        // Update metadata separately
-        // Math.maxでスプレッド演算子を使うとスタックオーバーフローになるため、reduceを使用
-        const lastTimestamp = allNewEvents.reduce((max, event) => {
-          const timestamp = event.timestamp || 0
-          return timestamp > max ? timestamp : max
-        }, 0)
-
-        await db.meta.put({
-          id: 'importStatus',
-          value: {
-            lastProcessedTimestamp: lastTimestamp,
-            lastProcessedEventCount: allNewEvents.length,
-            lastImportDate: new Date().toISOString()
-          },
-          updatedAt: Date.now()
-        })
-        console.log(`[importData] Updated metadata - lastTimestamp: ${lastTimestamp}`)
-
-        const entityTime = ((performance.now() - entityStartTime) / 1000).toFixed(2)
-        console.log(`[importData] Entity generation completed in ${entityTime}s`)
-
-      } catch (entityError) {
-        console.error('[importData] Entity generation error:', entityError)
-        // エラーの詳細をログに記録するが、処理は継続
-        // refreshDatabaseへのフォールバックは削除（トランザクション競合を避けるため）
-        const errorMessage = entityError instanceof Error ? entityError.message : String(entityError)
-        throw new Error(`Entity generation failed: ${errorMessage}`)
-      }
-    } else {
-      // 新規イベントがない場合は増分処理も不要
-      console.log('[importData] No new events to process')
-    }
-
-    // バッチモードを無効化
-    service.setBatchMode(false)
-
-    // インポート後に統計を強制的に更新
-    // 最新のEVT_DEALを取得して統計計算をトリガー
-    const latestDealEvent = await findLatestPlayerDealEvent(db)
-
-    if (latestDealEvent && isApiEventType(latestDealEvent, ApiType.EVT_DEAL)) {
-      // latestEvtDealを更新
-      service.latestEvtDeal = latestDealEvent
-
-      // プレイヤーIDも更新（インポートデータからヒーローを特定）
-      if (latestDealEvent.Player?.SeatIndex !== undefined) {
-        service.playerId = latestDealEvent.SeatUserIds[latestDealEvent.Player.SeatIndex]
-        console.log(`[importData] Updated playerId: ${service.playerId}`)
-      }
-
-      // 統計の再計算をトリガー
-      const playerIds = latestDealEvent.SeatUserIds.filter(id => id !== -1)
-      if (playerIds.length > 0) {
-        console.log('[importData] Triggering stats recalculation for imported data')
-        service.statsOutputStream.write(playerIds)
-
-        // 現在開いているゲームタブに対しても統計更新を通知
-        chrome.tabs.query({ url: gameUrlPattern }, tabs => {
-          tabs.forEach(tab => {
-            if (tab.id) {
-              chrome.tabs.sendMessage(tab.id, {
-                action: 'refreshStats'
-              })
-            }
-          })
-        })
-      }
-    }
-
-    currentOperationState = { type: 'idle' }
-    return { successCount, totalLines: lines.length, duplicateCount }
-
-  } catch (error) {
-    currentOperationState = { type: 'idle' }
-    console.error('Import error:', error)
-    throw error
-  }
-}
-
-/**
- * Service Worker のアイドル停止を防止するキープアライブを開始する。
- * Chrome MV3 では 30 秒のアイドル後に Worker が停止されるため、
- * 長時間のバッチ処理中は chrome.offscreen API でオフスクリーン
- * ドキュメントを作成して Worker を維持する。
- *
- * offscreen ドキュメントが存在する間、Worker は停止されない。
- * @returns クリーンアップ関数
- */
-const startKeepAlive = async (): Promise<() => void> => {
-  // offscreen API が利用可能な場合はそれを使用（Chrome 109+）
-  if (chrome.offscreen) {
-    try {
-      await chrome.offscreen.createDocument({
-        url: 'offscreen.html',
-        reasons: [chrome.offscreen.Reason.BLOBS],
-        justification: 'Keep service worker alive during batch export'
-      })
-    } catch (e) {
-      // 既に存在する場合は無視
-    }
-    return () => {
-      chrome.offscreen.closeDocument().catch(() => {})
-    }
-  }
-  
-  // フォールバック: setInterval + getPlatformInfo
-  const id = setInterval(() => {
-    chrome.runtime.getPlatformInfo().catch(() => {})
-  }, 25000)
-  return () => clearInterval(id)
-}
-
-const exportJsonData = async (db: PokerChaseDB) => {
-  const stopKeepAlive = await startKeepAlive()
-  try {
-    currentOperationState = { type: 'export', format: 'json', progress: 0 }
-    chrome.runtime.sendMessage<ExportProgressMessage>({
-      action: 'exportProgress',
-      state: 'started',
-      format: 'json',
-      message: 'NDJSONエクスポート開始...'
-    }).catch(() => {})
-
-    const totalCount = await db.apiEvents.count()
-    console.log(`[Export] Exporting ${totalCount} events...`)
-    
-    // Direct chunked export using primary key cursor to avoid Dexie Collection offset issues
-    const chunks: string[] = []
-    let processedCount = 0
-    let lastKey: any = undefined
-    const chunkSize = DATABASE_CONSTANTS.EXPORT_CHUNK_SIZE
-    
-    while (true) {
-      // Build fresh query each iteration using primary key range
-      const chunk = lastKey !== undefined
-        ? await db.apiEvents.where('[timestamp+ApiTypeId]').above(lastKey).limit(chunkSize).toArray()
-        : await db.apiEvents.orderBy('[timestamp+ApiTypeId]').limit(chunkSize).toArray()
-      
-      if (chunk.length === 0) break
-      
-      chunks.push(chunk.map(event => JSON.stringify(event)).join('\n'))
-      processedCount += chunk.length
-      
-      // Track last key for next iteration
-      const lastEvent = chunk[chunk.length - 1]!
-      lastKey = [lastEvent.timestamp, lastEvent.ApiTypeId]
-      
-      const progress = Math.round((processedCount / totalCount) * 100)
-      const progressMessage = `エクスポート中... ${processedCount.toLocaleString()}/${totalCount.toLocaleString()} (${progress}%)`
-      currentOperationState = { type: 'export', format: 'json', progress, processed: processedCount, total: totalCount, message: progressMessage }
-      chrome.runtime.sendMessage<ExportProgressMessage>({
-        action: 'exportProgress',
-        state: 'processing',
-        format: 'json',
-        progress,
-        processed: processedCount,
-        total: totalCount,
-        message: progressMessage
-      }).catch(() => {})
-
-      if (processedCount % 50000 === 0 || processedCount >= totalCount) {
-        console.log(`[Export] Processed ${processedCount}/${totalCount} events`)
-      }
-      
-      if (chunk.length < chunkSize) break // Last chunk
-    }
-    
-    const jsonlContent = chunks.join('\n')
-    
-    downloadFile(
-      jsonlContent,
-      'pokerchase_raw_data.ndjson',
-      'application/x-ndjson'
-    )
-    
-    console.log(`[Export] Export completed: ${processedCount} events (${(jsonlContent.length / 1024 / 1024).toFixed(1)}MB)`)
-    
-    currentOperationState = { type: 'idle' }
-    chrome.runtime.sendMessage<ExportProgressMessage>({
-      action: 'exportProgress',
-      state: 'completed',
-      format: 'json',
-      progress: 100,
-      processed: processedCount,
-      total: totalCount,
-      message: `NDJSONエクスポート完了: ${processedCount.toLocaleString()}件`
-    }).catch(() => {})
-    stopKeepAlive()
-  } catch (error) {
-    stopKeepAlive()
-    console.error('[Export] Export failed:', error)
-    currentOperationState = { type: 'idle' }
-    chrome.runtime.sendMessage<ExportProgressMessage>({
-      action: 'exportProgress',
-      state: 'error',
-      format: 'json',
-      message: `NDJSONエクスポート失敗: ${error}`
-    }).catch(() => {})
-    throw error
-  }
-}
-
-const exportPokerStarsData = async () => {
-  const stopKeepAlive = await startKeepAlive()
-  try {
-    currentOperationState = { type: 'export', format: 'pokerstars', progress: 0 }
-    chrome.runtime.sendMessage<ExportProgressMessage>({
-      action: 'exportProgress',
-      state: 'started',
-      format: 'pokerstars',
-      message: 'PokerStarsエクスポート開始...'
-    }).catch(() => {})
-
-    // Get the last session's hand history
-    const handHistory = await service.exportHandHistory(undefined, (processed, total) => {
-      const progress = Math.round((processed / total) * 100)
-      chrome.runtime.sendMessage<ExportProgressMessage>({
-        action: 'exportProgress',
-        state: 'processing',
-        format: 'pokerstars',
-        progress,
-        processed,
-        total,
-        message: `ハンドヒストリー変換中... ${processed.toLocaleString()}/${total.toLocaleString()} (${progress}%)`
-      }).catch(() => {})
-      currentOperationState = { type: 'export', format: 'pokerstars', progress, processed, total, message: `ハンドヒストリー変換中... ${processed}/${total} (${progress}%)` }
-    })
-
-    if (!handHistory) {
-      console.error('No hands found to export')
-      currentOperationState = { type: 'idle' }
-      chrome.runtime.sendMessage<ExportProgressMessage>({
-        action: 'exportProgress',
-        state: 'error',
-        format: 'pokerstars',
-        message: 'エクスポートするハンドが見つかりませんでした'
-      }).catch(() => {})
-      // Show notification to user
-      chrome.notifications.create({
-        type: 'basic',
-        iconUrl: chrome.runtime.getURL('icons/icon_48px.png'),
-        title: 'エクスポートエラー',
-        message: 'エクスポートするハンドが見つかりませんでした。ゲームをプレイしてから再度お試しください。'
-      })
-      return
-    }
-
-    downloadFile(
-      handHistory,
-      'pokerchase_hand_history.txt',
-      'text/plain'
-    )
-
-    stopKeepAlive()
-    currentOperationState = { type: 'idle' }
-    chrome.runtime.sendMessage<ExportProgressMessage>({
-      action: 'exportProgress',
-      state: 'completed',
-      format: 'pokerstars',
-      message: 'PokerStarsハンドヒストリーエクスポート完了'
-    }).catch(() => {})
-  } catch (error) {
-    stopKeepAlive()
-    console.error('Error exporting PokerStars format:', error)
-    currentOperationState = { type: 'idle' }
-    chrome.runtime.sendMessage<ExportProgressMessage>({
-      action: 'exportProgress',
-      state: 'error',
-      format: 'pokerstars',
-      message: `PokerStarsエクスポート失敗: ${error}`
-    }).catch(() => {})
-    // Show error notification
-    chrome.notifications.create({
-      type: 'basic',
-      iconUrl: chrome.runtime.getURL('icons/icon_48px.png'),
-      title: 'エクスポートエラー',
-      message: 'ハンドヒストリーのエクスポート中にエラーが発生しました。'
-    })
-    throw error
-  }
-}
-
-
-const downloadFile = (content: string, filename: string, contentType: string) => {
-  const timestamp = new Date().toISOString().replace(/[:.]/g, '-')
-
-  const extensionMatch = filename.match(/\.[^.]+$/)
-  const extension = extensionMatch ? extensionMatch[0] : ''
-
-  const baseFilename = extension ? filename.slice(0, -extension.length) : filename
-
-  const getFinalFilename = () => {
-    if (contentType.includes('ndjson') || filename.endsWith('.jsonl') || filename.endsWith('.ndjson')) {
-      return `${baseFilename}_${timestamp}.ndjson`
-    } else if (contentType.includes('json')) {
-      return `${baseFilename}_${timestamp}.json`
-    } else if (contentType.includes('text')) {
-      return `${baseFilename}_${timestamp}.txt`
-    } else {
-      return `${baseFilename}_${timestamp}${extension || '.dat'}`
-    }
-  }
-
-  const finalFilename = getFinalFilename()
-
-  // Send to content script for Blob-based download (avoids data URL size limits)
-  chrome.tabs.query({ url: gameUrlPattern }, async tabs => {
-    const tab = tabs.find(t => t.id)
-    if (tab?.id) {
-      const sizeMB = content.length / 1024 / 1024
-      const MAX_CHUNK_MB = 50 // Under Chrome's 64MiB message limit
-      const maxChunkSize = MAX_CHUNK_MB * 1024 * 1024
-
-      if (content.length <= maxChunkSize) {
-        chrome.tabs.sendMessage(tab.id, { action: 'downloadFile', content, filename: finalFilename, contentType })
-      } else {
-        // Split into chunks for large files
-        const totalChunks = Math.ceil(content.length / maxChunkSize)
-        console.log(`[Export] Splitting ${sizeMB.toFixed(1)}MB into ${totalChunks} chunks...`)
-        chrome.tabs.sendMessage(tab.id, { action: 'downloadFileInit', filename: finalFilename, contentType, totalChunks })
-        for (let i = 0; i < totalChunks; i++) {
-          const chunk = content.slice(i * maxChunkSize, (i + 1) * maxChunkSize)
-          chrome.tabs.sendMessage(tab.id, { action: 'downloadFileChunk', chunkIndex: i, chunk, totalChunks })
-        }
-        chrome.tabs.sendMessage(tab.id, { action: 'downloadFileFinish', filename: finalFilename, contentType })
-      }
-      console.log(`[Export] Download initiated via content script: ${finalFilename} (${sizeMB.toFixed(1)}MB)`)
-      return
-    }
-    // Fallback: data URL (may fail for large files >2MB)
-    console.warn('[Export] No game tab found, falling back to data URL download')
-    downloadViaDataUrl(content, finalFilename, contentType)
-  })
-}
-
-const downloadViaDataUrl = (content: string, finalFilename: string, contentType: string) => {
-  const base64Content = btoa(encodeURIComponent(content).replace(/%([0-9A-F]{2})/g, (_match, p1) => String.fromCharCode(parseInt(p1, 16))))
-  const dataUrl = `data:${contentType};base64,${base64Content}`
-
-  chrome.downloads.download({
-    url: dataUrl,
-    filename: finalFilename,
-    saveAs: true
-  }, () => {
-    if (chrome.runtime.lastError) {
-      console.error('Download error:', chrome.runtime.lastError)
-    }
-  })
-}
-
-
-/**
- * from `content_script.ts`
- * @see https://developer.chrome.com/docs/extensions/develop/concepts/messaging?hl=ja#port-lifetime
- * @see https://medium.com/@bhuvan.gandhi/chrome-extension-v3-mitigate-service-worker-timeout-issue-in-the-easiest-way-fccc01877abd
- *
- * メッセージの型保証:
- * 1. web_accessible_resource.ts: ApiTypeIdを持つメッセージのみ送信
- * 2. content_script.ts: ApiMessage型として受信・転送
- * 3. background.ts: ApiMessage型として受信し、完全な検証を実施
- */
-// 以下の3つのストリームリスナーは、Service Workerのライフタイム中に一度だけ登録する
-// （`onConnect`のたびに登録すると、ページリロード/再接続のたびにリスナーが積み重なり、
-// 特に`handLogStream`のハンドラーはNタブ分`chrome.tabs.sendMessage`を重複送信してしまう）
-
-// Listen for real-time stats from dedicated stream
-service.realTimeStatsStream.on('data', (data: { handId?: number, stats: AllPlayersRealTimeStats, timestamp: number }) => {
-  latestRealTimeStats = data.stats
-
-  // Send update with real-time stats only
-  if (lastKnownStats && lastKnownStats.length > 0) {
-    broadcastMessage({
-      stats: lastKnownStats,
-      evtDeal: service.latestEvtDeal,
-      realTimeStats: latestRealTimeStats
-    })
-  }
-})
-service.statsOutputStream.on('data', async (hand: PlayerStats[]) => {
-  lastKnownStats = hand // Store for later use
-
-  // Real-time stats are now handled by RealTimeStatsStream
-  broadcastMessage({
-    stats: hand,
-    evtDeal: service.latestEvtDeal,  // Include EVT_DEAL for seat mapping
-    realTimeStats: latestRealTimeStats  // Include latest real-time stats from stream
-  })
-})
-
-// Handle hand log events
-service.handLogStream.on('data', (event: HandLogEvent) => {
-  // Send to all tabs with the game
-  chrome.tabs.query({ url: gameUrlPattern }, tabs => {
-    tabs.forEach(tab => {
-      if (tab.id) {
-        chrome.tabs.sendMessage<HandLogEventMessage>(tab.id, {
-          action: 'handLogEvent',
-          event: event
-        })
-      }
-    })
-  })
-})
-
-chrome.runtime.onConnect.addListener(port => {
-  if (port.name === PokerChaseService.POKER_CHASE_SERVICE_EVENT) {
-    connectedPorts.add(port)
-    port.onMessage.addListener(async (message: ApiMessage | { type: string }) => {
-      // キープアライブメッセージの処理
-      if (typeof message === 'object' && 'type' in message && message.type === 'keepalive') {
-        return // キープアライブは処理のみで、それ以上何もしない
-      }
-
-      // Ensure service is ready before processing messages
-      try {
-        await service.ready
-      } catch (err) {
-        console.error('[background] Service not ready:', err)
-        return
-      }
-
-      // 通常のAPIメッセージ処理
-      // Zodスキーマでパース（passthrough: 未知プロパティは保持）
-      const data = parseApiEvent(message as ApiMessage)
-
-      if (!data) {
-        // パース失敗 = 必須プロパティ欠損など破壊的変更の可能性
-        const validationResult = validateApiEvent(message as ApiMessage)
-        const errorDetails = validationResult.error ? getValidationError(validationResult.error) : null
-        console.warn(`[background] Schema validation failed (event dropped):\n  Errors: ${JSON.stringify(errorDetails, null, 2)}\n  Event: ${JSON.stringify(message, null, 2)}`)
-        return
-      }
-
-      // アプリケーション用のイベントかチェック
-      if (!isApplicationApiEvent(data)) {
-        // アプリケーションで使用しないApiTypeIdのイベントはDB保存しないが内容は記録
-        console.info(`[background] Non-application event (${data.ApiTypeId}): ${JSON.stringify(data)}`)
-        return
-      }
-
-      // ここでdataはApiEvent型（isApplicationApiEventで保証済み）
-      service.eventLogger(data, 'info')
-
-      // DB保存とストリーム処理
-      service.db.apiEvents.add(data)
-        .catch(err => console.error('[background] Failed to save event:', err))
-      service.handLogStream.write(data)
-      service.handAggregateStream.write(data)
-      service.realTimeStatsStream.write(data)
-
-      // Handle game session end for auto sync
-      if (data.ApiTypeId === ApiType.EVT_SESSION_RESULTS) {
-        autoSyncService.onGameSessionEnd().catch(err =>
-          console.error('[background] Auto sync on game end failed:', err)
-        )
-      }
-    })
-    // PINGは自ポートにのみ送信する（ブロードキャストしない）
-    const pingPort = (data: string) => {
-      try {
-        port.postMessage(data)
-      } catch (error: unknown) {
-        if (error instanceof Error) {
-          /** when `content_script` is inactive */
-          if (error.message === 'Attempting to use a disconnected port object')
-            clearInterval(intervalId)
-          else
-            console.error(error)
-        }
-      }
-    }
-    const intervalId = setInterval(() => { pingPort(`[PING] ${new Date().toISOString()}`) }, PING_INTERVAL_MS)
-
-    // Clean up when port disconnects
-    port.onDisconnect.addListener(() => {
-      // Keep lastKnownStats for page reloads - only clear interval
-      clearInterval(intervalId)
-      connectedPorts.delete(port)
-    })
-  }
-})
-
-/**
- * Delete all data (logs only, not configuration)
- */
-const deleteAllData = async (): Promise<void> => {
-  try {
-    // データベースを完全に削除
-    await db.delete()
-
-    // データベースの新しいインスタンスを確保するために拡張機能をリロード
-    chrome.runtime.reload()
-  } catch (error) {
-    console.error('Error deleting data:', error)
-    throw error
-  }
-}
-
-/**
- * Get the latest session stats from the last known data or database
- */
-const getLatestSessionStats = async (): Promise<PlayerStats[]> => {
-  // Return empty array - stats will be calculated when game starts
-  // This avoids showing stale data and ensures EVT_DEAL is available for seat mapping
-  return []
-}
-
-/**
- * Rebuild all data from apiEvents using batch processing
- * Similar to download sync processing to avoid multiple HUD updates
- */
-const rebuildAllData = async (): Promise<void> => {
-  try {
-    console.log('[rebuildAllData] Starting batch rebuild of all data...')
-    const startTime = performance.now()
-
-    currentOperationState = { type: 'rebuild', progress: 0, message: 'データ再構築開始...' }
-    chrome.runtime.sendMessage<RebuildProgressMessage>({
-      action: 'rebuildProgress',
-      state: 'started',
-      message: 'データ再構築開始...'
-    }).catch(() => {})
-
-    // Clear all entity tables first
-    await db.transaction('rw', [db.hands, db.phases, db.actions, db.meta], async () => {
-      await db.hands.clear()
-      await db.phases.clear()
-      await db.actions.clear()
-      await db.meta.delete('lastProcessed')
-    })
-
-    currentOperationState = { type: 'rebuild', progress: 10, message: 'テーブルクリア完了、イベント読み込み中...' }
-    chrome.runtime.sendMessage<RebuildProgressMessage>({
-      action: 'rebuildProgress',
-      state: 'processing',
-      progress: 10,
-      message: 'テーブルクリア完了、イベント読み込み中...'
-    }).catch(() => {})
-
-    // Get total event count
-    const totalCount = await db.apiEvents.count()
-    console.log(`[rebuildAllData] Processing ${totalCount} events...`)
-
-    if (totalCount === 0) {
-      console.log('[rebuildAllData] No events to process')
-      currentOperationState = { type: 'idle' }
-      chrome.runtime.sendMessage<RebuildProgressMessage>({
-        action: 'rebuildProgress',
-        state: 'completed',
-        progress: 100,
-        message: '処理対象のイベントがありません'
-      }).catch(() => {})
-      return
-    }
-
-    // Enable batch mode to prevent real-time updates
-    service.setBatchMode(true)
-
-    try {
-      // Process in chunks to avoid memory issues
-      let totalHands = 0
-      let totalPhases = 0
-      let totalActions = 0
-      
-      // Initialize EntityConverter
-      const defaultSession = {
-        id: undefined,
-        battleType: undefined,
-        name: undefined,
-        players: new Map(),
-        reset: () => { }
-      }
-      const converter = new EntityConverter(defaultSession)
-      
-      // Load all events and convert in one pass
-      // (EntityConverter tracks hand state internally, so chunked conversion loses cross-chunk hands)
-      console.log(`[rebuildAllData] Loading all events...`)
-      const allEvents = await db.apiEvents.orderBy('[timestamp+ApiTypeId]').toArray()
-      console.log(`[rebuildAllData] Loaded ${allEvents.length} events, converting to entities...`)
-      
-      currentOperationState = { type: 'rebuild', progress: 40, message: `${allEvents.length.toLocaleString()}件のイベントを変換中...` }
-      chrome.runtime.sendMessage<RebuildProgressMessage>({
-        action: 'rebuildProgress',
-        state: 'processing',
-        progress: 40,
-        message: `${allEvents.length.toLocaleString()}件のイベントを変換中...`
-      }).catch(() => {})
-
-      const entities = converter.convertEventsToEntities(allEvents)
-      
-      currentOperationState = { type: 'rebuild', progress: 70, message: 'エンティティ保存中...' }
-      chrome.runtime.sendMessage<RebuildProgressMessage>({
-        action: 'rebuildProgress',
-        state: 'processing',
-        progress: 70,
-        message: 'エンティティ保存中...'
-      }).catch(() => {})
-
-      const counts = await saveEntities(db, entities)
-      totalHands += counts.hands
-      totalPhases += counts.phases
-      totalActions += counts.actions
-      
-      console.log(`[rebuildAllData] Generated entities - Hands: ${totalHands}, Phases: ${totalPhases}, Actions: ${totalActions}`)
-
-      // Restore service state from latest events  
-      const latestDealEvent = await findLatestPlayerDealEvent(db)
-
-      if (latestDealEvent && isApiEventType(latestDealEvent, ApiType.EVT_DEAL)) {
-        service.latestEvtDeal = latestDealEvent
-        if (latestDealEvent.Player?.SeatIndex !== undefined) {
-          service.playerId = latestDealEvent.SeatUserIds[latestDealEvent.Player.SeatIndex]
-        }
-      }
-
-      // Update metadata with rebuild info
-      await db.meta.put({
-        id: 'rebuildStatus',
-        value: {
-          lastRebuildDate: new Date().toISOString(),
-          totalEvents: totalCount,
-          totalHands: totalHands,
-          totalPhases: totalPhases,
-          totalActions: totalActions
-        },
-        updatedAt: Date.now()
-      })
-
-      const rebuildTime = ((performance.now() - startTime) / 1000).toFixed(2)
-      console.log(`[rebuildAllData] Rebuild completed in ${rebuildTime}s`)
-
-      currentOperationState = { type: 'rebuild', progress: 90, message: '統計情報を再計算中...' }
-      chrome.runtime.sendMessage<RebuildProgressMessage>({
-        action: 'rebuildProgress',
-        state: 'processing',
-        progress: 90,
-        message: '統計情報を再計算中...'
-      }).catch(() => {})
-
-      // Trigger stats recalculation once at the end
-      if (service.latestEvtDeal && service.latestEvtDeal.SeatUserIds) {
-        const playerIds = service.latestEvtDeal.SeatUserIds.filter(id => id !== -1)
-        if (playerIds.length > 0) {
-          console.log('[rebuildAllData] Triggering stats recalculation...')
-          service.statsOutputStream.write(service.latestEvtDeal.SeatUserIds)
-        }
-      }
-
-      currentOperationState = { type: 'idle' }
-      chrome.runtime.sendMessage<RebuildProgressMessage>({
-        action: 'rebuildProgress',
-        state: 'completed',
-        progress: 100,
-        message: `データ再構築完了 (${rebuildTime}秒) - ハンド: ${totalHands.toLocaleString()}, フェーズ: ${totalPhases.toLocaleString()}, アクション: ${totalActions.toLocaleString()}`
-      }).catch(() => {})
-
-    } finally {
-      // Disable batch mode
-      service.setBatchMode(false)
-    }
-
-  } catch (error) {
-    console.error('[rebuildAllData] Error:', error)
-    currentOperationState = { type: 'idle' }
-    chrome.runtime.sendMessage<RebuildProgressMessage>({
-      action: 'rebuildProgress',
-      state: 'error',
-      message: `データ再構築失敗: ${error}`
-    }).catch(() => {})
-    throw error
-  }
-}
-
-/**
- * Firebase Auth Handlers
- */
-const handleFirebaseSignIn = async (): Promise<void> => {
-  try {
-    const user = await firebaseAuthService.signInWithGoogle()
-    console.log('[Firebase] User signed in:', user.email)
-
-    // Initialize auto sync after sign in
-    await autoSyncService.onAuthStateChanged(user)
-  } catch (error) {
-    console.error('[Firebase] Sign in error:', error)
-    throw error
-  }
-}
-
-const handleFirebaseSignOut = async (): Promise<void> => {
-  try {
-    await firebaseAuthService.signOut()
-    console.log('[Firebase] User signed out')
-
-    // Update sync state
-    await autoSyncService.onAuthStateChanged(null)
-  } catch (error) {
-    console.error('[Firebase] Sign out error:', error)
-    throw error
-  }
-}
-
-
+registerEventIngestion(service)
 
 // Listen for auth state changes on startup
 firebaseAuthService.onAuthStateChange((user) => {
@@ -1326,4 +119,3 @@ firebaseAuthService.onAuthStateChange((user) => {
     : { isSignedIn: false, userInfo: null }
   chrome.storage.local.set({ firebaseAuthCache: authCache })
 })
-
