+++ src/background/ports.ts
@@ -1,0 +1,123 @@
+/** !!! CONTENT_SCRIPTS、WEB_ACCESSIBLE_RESOURCESからインポートしないこと !!! */
+import type PokerChaseService from '../services/poker-chase-service'
+import type { ApiEvent, ApiType, PlayerStats } from '../app'
+import type { AllPlayersRealTimeStats } from '../realtime-stats/realtime-stats-service'
+import type { HandLogEvent } from '../types/hand-log'
+import type { HandLogEventMessage } from '../types/messages'
+
+const PING_INTERVAL_MS = 10 * 1000
+
+let lastKnownStats: PlayerStats[] = []
+
+// Store latest real-time stats (module scope: shared across all port connections)
+let latestRealTimeStats: AllPlayersRealTimeStats | undefined
+
+export const getLastKnownStats = (): PlayerStats[] => lastKnownStats
+
+export const setLastKnownStats = (stats: PlayerStats[]): void => {
+  lastKnownStats = stats
+}
+
+export const getLatestRealTimeStats = (): AllPlayersRealTimeStats | undefined => latestRealTimeStats
+
+/**
+ * `chrome.runtime.onConnect`で接続されたポートの集合
+ * ストリームイベントをブロードキャストする際の送信先として利用する
+ */
+export const connectedPorts = new Set<chrome.runtime.Port>()
+
+/**
+ * 接続中の全ポートにメッセージをブロードキャストする
+ * 切断済みポートを検出した場合は`connectedPorts`から取り除く
+ */
+export const broadcastMessage = (data: { stats: PlayerStats[], evtDeal?: ApiEvent<ApiType.EVT_DEAL>, realTimeStats?: AllPlayersRealTimeStats } | string) => {
+  connectedPorts.forEach(port => {
+    try {
+      port.postMessage(data)
+    } catch (error: unknown) {
+      if (error instanceof Error) {
+        /** when `content_script` is inactive */
+        if (error.message === 'Attempting to use a disconnected port object')
+          connectedPorts.delete(port)
+        else
+          console.error(error)
+      }
+    }
+  })
+}
+
+/**
+ * PINGは自ポートにのみ送信する（ブロードキャストしない）
+ * `chrome.runtime.onConnect`のポート接続時に呼び出し、クリーンアップ関数を受け取る
+ */
+export const startPortPing = (port: chrome.runtime.Port): () => void => {
+  const pingPort = (data: string) => {
+    try {
+      port.postMessage(data)
+    } catch (error: unknown) {
+      if (error instanceof Error) {
+        /** when `content_script` is inactive */
+        if (error.message === 'Attempting to use a disconnected port object')
+          clearInterval(intervalId)
+        else
+          console.error(error)
+      }
+    }
+  }
+  const intervalId = setInterval(() => { pingPort(`[PING] ${new Date().toISOString()}`) }, PING_INTERVAL_MS)
+  return () => clearInterval(intervalId)
+}
+
+/**
+ * from `content_script.ts`
+ * @see https://developer.chrome.com/docs/extensions/develop/concepts/messaging?hl=ja#port-lifetime
+ * @see https://medium.com/@bhuvan.gandhi/chrome-extension-v3-mitigate-service-worker-timeout-issue-in-the-easiest-way-fccc01877abd
+ *
+ * メッセージの型保証:
+ * 1. web_accessible_resource.ts: ApiTypeIdを持つメッセージのみ送信
+ * 2. content_script.ts: ApiMessage型として受信・転送
+ * 3. background.ts: ApiMessage型として受信し、完全な検証を実施
+ */
+// 以下の3つのストリームリスナーは、Service Workerのライフタイム中に一度だけ登録する
+// （`onConnect`のたびに登録すると、ページリロード/再接続のたびにリスナーが積み重なり、
+// 特に`handLogStream`のハンドラーはNタブ分`chrome.tabs.sendMessage`を重複送信してしまう）
+export const registerStreamSubscriptions = (service: PokerChaseService, gameUrlPattern: string): void => {
+  // Listen for real-time stats from dedicated stream
+  service.realTimeStatsStream.on('data', (data: { handId?: number, stats: AllPlayersRealTimeStats, timestamp: number }) => {
+    latestRealTimeStats = data.stats
+
+    // Send update with real-time stats only
+    if (lastKnownStats && lastKnownStats.length > 0) {
+      broadcastMessage({
+        stats: lastKnownStats,
+        evtDeal: service.latestEvtDeal,
+        realTimeStats: latestRealTimeStats
+      })
+    }
+  })
+  service.statsOutputStream.on('data', async (hand: PlayerStats[]) => {
+    lastKnownStats = hand // Store for later use
+
+    // Real-time stats are now handled by RealTimeStatsStream
+    broadcastMessage({
+      stats: hand,
+      evtDeal: service.latestEvtDeal,  // Include EVT_DEAL for seat mapping
+      realTimeStats: latestRealTimeStats  // Include latest real-time stats from stream
+    })
+  })
+
+  // Handle hand log events
+  service.handLogStream.on('data', (event: HandLogEvent) => {
+    // Send to all tabs with the game
+    chrome.tabs.query({ url: gameUrlPattern }, tabs => {
+      tabs.forEach(tab => {
+        if (tab.id) {
+          chrome.tabs.sendMessage<HandLogEventMessage>(tab.id, {
+            action: 'handLogEvent',
+            event: event
+          })
+        }
+      })
+    })
+  })
+}
